Fixed compilation issues

+++ Rendering/Light/AreaLight.hpp
@@ -18,7 +18,7 @@ namespace Odysseus
 
         virtual void setOrderOfExecution(const short& newOrderOfExecution);
 
-        virtual short getUniqueID();
+        virtual int getUniqueID();
 
         virtual std::string toString();
         void addLight(PointLight* pt);
