+++ Rendering/Light/SpotLight.hpp
@@ -24,7 +24,7 @@ namespace Odysseus
 
             virtual void setOrderOfExecution(const short& newOrderOfExecution);
 
-            virtual short getUniqueID();
+            virtual int getUniqueID();
 
             virtual std::string toString();
